+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - [pull #590] Fix underscores within bold text getting emphasized (#589)
 - [pull #591] Add Alerts extra
+- [pull #595] Fix img alt text being processed as markdown (#594)
 
 
 ## python-markdown2 2.5.0